requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
cffi==1.14.5
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.0.4
chasgimqtt==0.3.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.7
daphne==2.5.0
defusedxml==0.7.1
Django==3.2
django-allauth==0.45.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
kombu==5.0.2
msgpack==0.6.2
oauthlib==3.1.1
paho-mqtt==1.5.1
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
